services/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
google-generativeai==0.8.3
ollama==0.1.6
python-dotenv==1.0.0
pydantic==2.5.3
